refactor: merge PopularHook to PegasusHook
